requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
Django==4.1.5
django-autoslug==1.9.8
django-bootstrap4==22.3
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-etc==1.4.0
django-hitcount==1.3.5
django-js-asset==2.0.0
Pillow==9.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
